Dungeons

+++ Assets/Scripts/BigMap/BigMapInteraction/Store.cs
@@ -14,15 +14,21 @@ public class Store : InteractionObject {
 
     protected override string GetName() {
         Vector3 pos = this.transform.position;
-        return $"Store_{pos.x}_{pos.y}_{pos.z}";
+        string dungeonName = SceneChangeManager.Instance.CurrentDungeonName;
+        return $"{dungeonName}_Store_{pos.x}_{pos.y}_{pos.z}";
     }
 
     protected override void LoadBigMapData() {
-        if (!this.IsEnd) {
-            if(PlayerPrefs.HasKey(GetName())) PlayerPrefs.DeleteKey(GetName());
-            this.EndInteraction();
+        if (this.Goods == null || this.Goods.Count == 0) {
+            // Random Store
+            if (!this.IsEnd) {
+                if(PlayerPrefs.HasKey(GetName())) PlayerPrefs.DeleteKey(GetName());
+                this.EndInteraction();
+                
+                // TODO: Random Store Goods
+            }
         }
-
+        
         if (PlayerPrefs.HasKey(GetName())) {
             string json = PlayerPrefs.GetString(GetName());
             this.Goods = JsonUtility.FromJson<Serialization<string>>(json).ToList();
